Refactor rendering of index view
Use Django's 'render' instead of 'loader' to apply the index.html template. It rolls the loading of a template (with context) and the formation of an HTTP response into a single function call. When the rest of the placeholder views have been improved to use templates, the views file will no longer need to import HttpResponse.